+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## 0.6.0-SNAPSHOT
+
+* **BREAKING**: [Cheshire](https://github.com/dakrone/cheshire) in dropped in favor of [Jsonista](https://github.com/metosin/jsonista) as the default JSON formatter (much faster, explicit configuration)
+  * `muuntaja.format.json` => `muuntaja.format.cheshire`
+  * `muuntaja.format.jsonista` => `muuntaja.format.json`
+  * The `muuntaja.format.json` formatter parameters have changed, with a asertion to fail fast:
+     * `:key-fn` => `:encode-key-fn` and `:decode-key-fn`
+     * `:bigdecimals?` => `:bigdecimals`
+
+* dropped dependencies:
+
+```clj
+[cheshire "5.8.0"]
+```
+
+* added dependencies:
+
+```clj
+[metosin/jsonista "0.2.1-SNAPSHOT"]
+```
+
 ## 0.5.0 (17.1.2018)
 
 * Fix [Cannot use muuntaja.core without depending on Ring](https://github.com/metosin/muuntaja/issues/58).